Assignment 4
Assignment 4 : Develop a mobile application that takes the user input and sends it to IoT device (python code). print the received data in python shell.
Keep a text box to accept the user input.integrate a submit button.
whenever user enters the text input in text box and clicks the button the data should be sent to IBM cloud using URL(HTTP API).